backend/requirements.txt
twilio==8.10.0
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.7
pydantic==2.5.3
pydantic-settings==2.1.0
scikit-learn==1.4.0
xgboost==2.0.3
tensorflow==2.15.0
keras==2.15.0
shap==0.44.0
joblib==1.3.2
numpy==1.26.3
pandas==2.1.4
geopandas==0.14.2
shapely==2.0.2
folium==0.15.1
pytesseract==0.3.10
Pillow==10.2.0
spacy==3.7.2
transformers==4.37.0
networkx==3.2.1
statsmodels==0.14.1
reportlab==4.0.9
openpyxl==3.1.2
matplotlib==3.8.2
seaborn==0.13.1
slowapi==0.1.9
httpx==0.26.0
pytest==7.4.4
pytest-asyncio==0.23.3
python-dotenv==1.0.0
scipy==1.12.0
